load/save implementation #1 fixed a warning msg in widget.cpp added a new KEY_DELETE in the list of available keys

+++ freesynd/trunk/src/menus/widget.cpp
@@ -389,7 +389,7 @@ void TeamListBox::handleMouseMotion(int x, int y, int state, const int modKeys) 
     if (pModel_) {
         // Gets the line pointed by the mouse
         int i = (y - yOrigin_) / 12;
-        if (i >= 0 && i < pModel_->size()) {
+        if (i >= 0 && i < (int) pModel_->size()) {
             if (pModel_->getElement(i)) {
                 // If line contains something, highlight it
                 if (focusedLine_ != i) {
